requirements.txt
asgiref==3.7.2
Django==4.2.6
pip==23.2.1
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3
wheel==0.41.2
djangorestframework==3.14.0